Merged PR #8: Support audio and music
This PR support audio and sound partially. Right now, we are able to load new sounds to Sound objects and play them. One thing to note that is we need to configure the debugging location to DirectXGame project so that the codes can understand resource locations.

Moreover, this PR turned off conformance mode to bypass C7510 error. Reference: https://stackoverflow.com/questions/51864528/update-visual-studio-2017-now-getting-compile-error-c7510-callback-use-of-d